test - readOnly distortions should have override fix
