+++ test/attacks.js
@@ -1,6 +1,9 @@
 const test = require('tape-promise').default(require('tape'))
 
-const { createBundleFromRequiresArray } = require('./util')
+const {
+  testEntryAttackerVictim,
+  createBundleFromRequiresArray
+} = require('./util')
 
 test('attack - prevent primitive modification', async (t) => {
   const depsArray = [
@@ -100,58 +103,3 @@ test('attack - prevent module cache attack in proto chain', async (t) => {
 
   await testEntryAttackerVictim(t, { defineAttacker, defineVictim })
 })
-
-async function testEntryAttackerVictim (t, { defineAttacker, defineVictim }) {
-
-  function defineEntry () {
-    require('attacker')
-    const result = require('victim').action()
-    global.testResult = result
-  }
-
-  const depsArray = [
-    {
-      'id': '/entry.js',
-      'file': '/entry.js',
-      'source': `(${defineEntry})()`,
-      'deps': {
-        'attacker': '/node_modules/attacker/index.js',
-        'victim': '/node_modules/victim/index.js'
-      },
-      'entry': true
-    },
-    {
-      'id': '/node_modules/attacker/index.js',
-      'file': '/node_modules/attacker/index.js',
-      'source': `(${defineAttacker})()`,
-      'deps': {
-        'victim': '/node_modules/victim/index.js'
-      }
-    },
-    {
-      'id': '/node_modules/victim/index.js',
-      'file': '/node_modules/victim/index.js',
-      'source': `(${defineVictim})()`,
-      'deps': {}
-    }
-  ]
-
-  const config = {
-    "resources": {
-      "<root>": {
-        "packages": {
-          "attacker": true,
-          "victim": true,
-        }
-      },
-      "attacker": {
-        "packages": {
-          "victim": true,
-        }
-      },
-    }
-  }
-  const result = await createBundleFromRequiresArray(depsArray, { lavamoatConfig: config })
-  eval(result)
-  t.equal(global.testResult, false)
-}
